# src/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "API de Filmes TMDB",
      "description": "API para gerenciar filmes consumindo a API pública TMDB"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": ["http"],
    "paths": {
      "/filme": {
        "post": {
          "summary": "Adiciona um filme à lista de desejos",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "nome": { "type": "string", "example": "Carros" }
                }
              }
            }
          ],
          "responses": {
            "200": { "description": "Filme adicionado com sucesso" },
            "404": { "description": "Filme não encontrado na API externa" }
          }
        },
        "get": {
          "summary": "Lista todos os filmes na lista de desejos",
          "responses": {
            "200": {
              "description": "Lista de filmes",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Filme"
                }
              }
            }
          }
        }
      },
      "/filme/{id}": {
        "get": {
          "summary": "Retorna detalhes de um filme específico",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Detalhes do filme",
              "schema": { "$ref": "#/definitions/Filme" }
            },
            "404": { "description": "Filme não encontrado" }
          }
        }
      },
...
    },
    "definitions": {
      "Filme": {
        "type": "object",
        "properties": {
          "uuid": { "type": "string", "example": "123e4567-e89b-12d3-a456-426614174000" },
          "titulo": { "type": "string", "example": "Clube da Luta" },
          "sinopse": { "type": "string", "example": "Um homem insone conhece um vendedor de sabão peculiar e forma um clube secreto de luta." },
          "ano_lancamento": { "type": "integer", "example": 1999 },
          "genero": { "type": "string", "example": "Drama" },
          "estado": { "type": "string", "example": "para assistir" }
        }
      },
...
    }
  }